Rename `XTX_Extractor` to `SwitchSwizzle` and move the texture converter to it
