+++ JustDanceEditor.Converter/Converters/Bundles/CoachesLargeBundleGenerator.cs
@@ -3,10 +3,11 @@ using AssetsTools.NET.Extra;
 using AssetsTools.NET.Texture;
 
 using JustDanceEditor.Converter.Unity;
-using JustDanceEditor.Converter.Unity.TextureConverter;
 
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.PixelFormats;
+
+using SwitchTexture.TextureConverter;
 
 namespace JustDanceEditor.Converter.Converters.Bundles;
