+++ JustDanceEditor.Converter/Converters/Bundles/CoachesSmallBundleGenerator.cs
@@ -3,11 +3,12 @@ using AssetsTools.NET.Texture;
 using AssetsTools.NET;
 
 using JustDanceEditor.Converter.Unity;
-using JustDanceEditor.Converter.Unity.TextureConverter;
 
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.PixelFormats;
 using SixLabors.ImageSharp.Processing;
+
+using SwitchTexture.TextureConverter;
 
 namespace JustDanceEditor.Converter.Converters.Bundles;
