+++ SwitchTexture/SwitchTexture.csproj
@@ -1,0 +1,25 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+	<PropertyGroup>
+		<OutputType>Exe</OutputType>
+		<TargetFramework>net8.0</TargetFramework>
+		<ImplicitUsings>enable</ImplicitUsings>
+		<Nullable>enable</Nullable>
+		<AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+	</PropertyGroup>
+
+	<ItemGroup>
+	  <PackageReference Include="AssetsTools.NET.Texture" Version="1.0.0" />
+	  <PackageReference Include="SixLabors.ImageSharp" Version="3.1.5" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <ProjectReference Include="..\TexToolWrap\TexToolWrap.vcxproj" />
+	</ItemGroup>
+
+	<Target Name="CopyFilesAfterBuild" AfterTargets="AfterBuild">
+		<Copy SourceFiles="$(SolutionDir)TexToolWrap\PVRTexLib\Windows_x86_64\PVRTexLib.dll" DestinationFolder="$(TargetDir)" />
+		<Copy SourceFiles="$(SolutionDir)TexToolWrap\ispc\win64\ispc_texcomp.dll" DestinationFolder="$(TargetDir)" />
+		<Copy SourceFiles="$(SolutionDir)TexToolWrap\crunch\win64\crnlib.dll" DestinationFolder="$(TargetDir)" />
+	</Target>
+</Project>
