--- XTX_Extractor/Swizzle.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace XTX_Extractor;
-internal class Swizzle
-{
-    internal static byte[] Deswizzle(uint v1, uint v2, XTX.XTXImageFormat format, byte[] mipData) => throw new NotImplementedException();
-}
